app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
reportlab==4.2.2
xhtml2pdf==0.2.13
Flask-Mail==0.10.0